fix(theming, modals, colorpickers): refactor types and interfaces (#1326)
