+++ packages/colorpickers/src/elements/ColorpickerDialog/index.tsx
@@ -12,74 +12,21 @@ import React, {
   Children,
   cloneElement,
   forwardRef,
-  ReactElement,
-  HTMLAttributes
+  ReactElement
 } from 'react';
 import PropTypes from 'prop-types';
-import { Modifier } from 'react-popper';
 import { Button } from '@zendeskgarden/react-buttons';
-import { GARDEN_PLACEMENT } from '@zendeskgarden/react-modals';
+import { PLACEMENT } from '@zendeskgarden/react-modals';
 import { composeEventHandlers } from '@zendeskgarden/container-utilities';
 import Chevron from '@zendeskgarden/svg-icons/src/16/chevron-down-stroke.svg';
-import { Colorpicker, IColorpickerProps } from '../Colorpicker';
-import { IColor } from '../../utils/types';
+import { Colorpicker } from '../Colorpicker';
 import {
   StyledButton,
   StyledButtonPreview,
   StyledTooltipModal,
   StyledTooltipBody
 } from '../../styled';
-
-export interface IColorpickerDialogChanges {
-  isOpen?: boolean;
-}
-
-export interface IColorpickerDialogProps extends IColorpickerProps {
-  /**
-   * Handles close actions. Can be triggered from the backdrop.
-   *
-   * @param {Object} color A color picker state
-   */
-  onClose?: (color: IColor) => void;
-  /** Adjusts the placement of the color dialog */
-  placement?: GARDEN_PLACEMENT;
-  /** Disables the color dialog button */
-  disabled?: boolean;
-  /**
-   * Modifies [Popper instance](https://popper.js.org/docs/v2/modifiers/) to customize positioning logic
-   */
-  popperModifiers?: Partial<Modifier<any, any>>[];
-  /**
-   * Sets the `z-index` of the color dialog
-   */
-  zIndex?: number;
-  /**
-   * Adds an arrow to the color dialog
-   */
-  hasArrow?: boolean;
-  /**
-   * Animates the color dialog
-   */
-  isAnimated?: boolean;
-  /**
-   * Opens the dialog in a controlled color picker dialog
-   */
-  isOpen?: boolean;
-  /**
-   * Applies inset `box-shadow` styling on focus
-   */
-  focusInset?: boolean;
-  /**
-   * Passes HTML attributes to the color dialog button element
-   */
-  buttonProps?: HTMLAttributes<HTMLButtonElement>;
-  /**
-   * Handles dialog changes
-   *
-   * @param {Object} changes The changed dialog state
-   */
-  onDialogChange?: (changes: IColorpickerDialogChanges) => void;
-}
+import { IColor, IColorpickerDialogProps } from '../../types';
 
 /**
  * @extends HTMLAttributes<HTMLDivElement>
@@ -201,29 +148,19 @@ export const ColorpickerDialog = forwardRef<HTMLDivElement, IColorpickerDialogPr
 );
 
 ColorpickerDialog.propTypes = {
-  placement: PropTypes.oneOf([
-    'auto',
-    'top',
-    'top-start',
-    'top-end',
-    'end',
-    'end-top',
-    'end-bottom',
-    'bottom',
-    'bottom-start',
-    'bottom-end',
-    'start',
-    'start-top',
-    'start-bottom'
-  ]),
+  ...Colorpicker.propTypes,
+  placement: PropTypes.oneOf(PLACEMENT),
   onClose: PropTypes.func,
-  onChange: PropTypes.func,
   onDialogChange: PropTypes.func,
   disabled: PropTypes.bool,
   labels: PropTypes.object,
-  color: PropTypes.oneOfType<any>([PropTypes.object, PropTypes.string]),
-  defaultColor: PropTypes.oneOfType<any>([PropTypes.object, PropTypes.string]),
-  buttonProps: PropTypes.object
+  buttonProps: PropTypes.object,
+  popperModifiers: PropTypes.any,
+  zIndex: PropTypes.number,
+  hasArrow: PropTypes.bool,
+  isAnimated: PropTypes.bool,
+  isOpen: PropTypes.bool,
+  focusInset: PropTypes.bool
 };
 
 ColorpickerDialog.defaultProps = {
