+++ packages/colorpickers/src/index.ts
@@ -5,15 +5,16 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-export type { IColor } from './utils/types';
 export { Colorpicker } from './elements/Colorpicker';
-export type { IColorpickerProps } from './elements/Colorpicker';
 export { ColorpickerDialog } from './elements/ColorpickerDialog';
-export type {
-  IColorpickerDialogProps,
-  IColorpickerDialogChanges
-} from './elements/ColorpickerDialog';
 export { ColorSwatch } from './elements/ColorSwatch';
-export type { IColorSwatchProps, ILabeledColor } from './elements/ColorSwatch';
 export { ColorSwatchDialog } from './elements/ColorSwatchDialog';
-export type { IColorSwatchDialogProps } from './elements/ColorSwatchDialog';
+
+export type {
+  IColor,
+  IColorpickerProps,
+  IColorpickerDialogProps,
+  ILabeledColor,
+  IColorSwatchProps,
+  IColorSwatchDialogProps
+} from './types';
