+++ packages/modals/.size-snapshot.json
@@ -1,20 +1,20 @@
 {
   "index.cjs.js": {
-    "bundled": 54058,
-    "minified": 38670,
-    "gzipped": 8244
+    "bundled": 54125,
+    "minified": 38739,
+    "gzipped": 8287
   },
   "index.esm.js": {
-    "bundled": 50262,
-    "minified": 35452,
-    "gzipped": 7993,
+    "bundled": 50369,
+    "minified": 35554,
+    "gzipped": 8031,
     "treeshaked": {
       "rollup": {
-        "code": 28091,
+        "code": 28055,
         "import_statements": 744
       },
       "webpack": {
-        "code": 31465
+        "code": 31423
       }
     }
   }
