+++ packages/modals/src/index.ts
@@ -6,7 +6,6 @@
  */
 
 export { Modal } from './elements/Modal';
-export type { IModalProps } from './elements/Modal';
 
 export { Body } from './elements/Body';
 export { Close } from './elements/Close';
@@ -15,9 +14,15 @@ export { FooterItem } from './elements/FooterItem';
 export { Header } from './elements/Header';
 
 export { TooltipModal } from './elements/TooltipModal/TooltipModal';
-export type { ITooltipModalProps } from './elements/TooltipModal/TooltipModal';
 
 export { DrawerModal } from './elements/DrawerModal/DrawerModal';
-export type { IDrawerModalProps } from './elements/DrawerModal/DrawerModal';
 
-export type { GARDEN_PLACEMENT } from './utils/gardenPlacements';
+export { PLACEMENT } from './types';
+
+export type {
+  IModalProps,
+  IDrawerModalProps,
+  ITooltipModalProps,
+  /* @deprecated type can be dereferenced from the exported interfaces */
+  Placement as GARDEN_PLACEMENT
+} from './types';
