+++ packages/theming/demo/utilities.stories.mdx
@@ -1,10 +1,15 @@
 import { Meta, ArgsTable, Canvas, Story } from '@storybook/addon-docs';
+import { ThemeProvider, DEFAULT_THEME, PALETTE } from '@zendeskgarden/react-theming';
 import { ArrowStylesStory } from './stories/ArrowStylesStory';
 import { MenuStylesStory } from './stories/MenuStylesStory';
 import { GetColorStory } from './stories/GetColorStory';
 import { ARROW_POSITIONS, MENU_POSITIONS } from './stories/data';
 
-<Meta title="Packages/Theming" />
+<Meta
+  title="Packages/Theming"
+  component={ThemeProvider}
+  subcomponents={{ DEFAULT_THEME, PALETTE }}
+/>
 
 # API
