+++ packages/theming/src/index.ts
@@ -6,9 +6,7 @@
  */
 
 export { ThemeProvider } from './elements/ThemeProvider';
-export type { IThemeProviderProps } from './elements/ThemeProvider';
 export { default as DEFAULT_THEME } from './elements/theme';
-export type { IGardenTheme } from './elements/theme';
 export { default as PALETTE } from './elements/palette';
 export { default as isRtl } from './utils/isRtl';
 export {
@@ -22,7 +20,17 @@ export { default as getColor } from './utils/getColor';
 export { default as getLineHeight } from './utils/getLineHeight';
 export { default as mediaQuery } from './utils/mediaQuery';
 export { default as arrowStyles } from './utils/arrowStyles';
-export type { ARROW_POSITION } from './utils/arrowStyles';
 export { useDocument } from './utils/useDocument';
 export { default as menuStyles } from './utils/menuStyles';
-export type { MENU_POSITION } from './utils/menuStyles';
+
+export {
+  ARROW_POSITION as ARRAY_ARROW_POSITION,
+  MENU_POSITION as ARRAY_MENU_POSITION
+} from './types';
+
+export type {
+  IGardenTheme,
+  IThemeProviderProps,
+  ArrowPosition as ARROW_POSITION,
+  MenuPosition as MENU_POSITION
+} from './types';
